:sparkles: include starter files to directory for components, work on blog layout
the blog layout uses grid template areas to assign sections for the blog. we assume a blog has three sections - aside-content-aside; like typical blogs found on the internet. the current tasks remaining for the blog layout is to add cards and reusable components that are largely used in blogs like profile, span images, floating images, etc.